完成情緒網站

+++ src/pages/MyPosts.js
@@ -1,5 +1,5 @@
 import { Item, Header, Button } from "semantic-ui-react";
-import { useEffect, useState } from "react";
+import { Fragment, useEffect, useState } from "react";
 import firebase from "../utils/firebase";
 import Post from "../components/Post";
 
@@ -29,6 +29,12 @@ function MyPosts() {
         try {
             const docRef = firebase.firestore().collection("posts").doc(postId);
             console.log('Document reference:', docRef);
+            const confirmed = window.confirm("確定要刪除文章嗎?");
+            if (!confirmed) {
+                console.log("刪除已取消");
+                return;
+            }
+
             await docRef.delete();
             console.log('Post deleted successfully');
             setPosts(posts.filter(post => post.id !== postId));
@@ -43,10 +49,10 @@ function MyPosts() {
         <Item.Group>
             {posts.map(post => {
                 return (
-                    <>
-                    <Post post={post} key={post.id}/>
+                    <Fragment key={post.id}>
+                    <Post post={post} />
                     <Button negative onClick={() => deletePost(post.id)}>刪除文章</Button>
-                    </>
+                    </Fragment>
                 )
             })}
         </Item.Group>
